bug: 依赖job重试成功后导致矩阵插件状态异常 #8523
